resultsUI/requirements1.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
alembic=1.13.2=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.32.0=pypi_0
anyio=4.4.0=pypi_0
asgiref=3.8.1=pypi_0
astunparse=1.6.3=pypi_0
attrs=23.2.0=pypi_0
audeer=2.1.1=pypi_0
audformat=1.3.0=pypi_0
audinterface=1.2.2=pypi_0
audiofile=1.4.0=pypi_0
audioread=3.0.1=pypi_0
audmath=1.4.1=pypi_0
audobject=0.7.11=pypi_0
audresample=1.3.3=pypi_0
av=12.3.0=pypi_0
banal=1.0.6=pypi_0
beautifulsoup4=4.12.3=pypi_0
bitsandbytes=0.43.1=pypi_0
blinker=1.8.2=pypi_0
blobfile=2.1.1=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.7.4=h56e8100_0
celluloid=0.2.0=pypi_0
certifi=2024.7.4=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cmake=3.30.1=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.2.1=pypi_0
cuda-version=11.8=h70ddcb2_3
cudatoolkit=11.8.0=h09e9e62_13
cudnn=8.9.7.29=he6de189_3
cycler=0.12.1=pypi_0
decorator=4.4.2=pypi_0
deepface=0.0.92=pypi_0
distro=1.9.0=pypi_0
django=4.2.14=pypi_0
djangorestframework=3.15.2=pypi_0
djangorestframework-simplejwt=5.3.1=pypi_0
dlib=19.24.4=cuda118py39h319e9ad_1
dnspython=2.6.1=pypi_0
easing-functions=1.0.4=pypi_0
easydict=1.13=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.2.2=pypi_0
fairscale=0.4.13=pypi_0
fastapi=0.111.1=pypi_0
fastapi-cli=0.0.4=pypi_0
ffmpeg=1.4=pypi_0
ffmpy=0.4.0=pypi_0
filelock=3.15.4=pypi_0
fire=0.6.0=pypi_0
flask=3.0.3=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.6.1=pypi_0
gast=0.6.0=pypi_0
gdown=5.2.0=pypi_0
google-pasta=0.2.0=pypi_0
gradio=4.39.0=pypi_0
gradio-client=1.1.1=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.65.1=pypi_0
gunicorn=22.0.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.11.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.24.0=pypi_0
idna=3.7=pypi_0
imageio=2.34.2=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2021.4.0=pypi_0
iso-639=0.4.5=pypi_0
iso3166=2.1.1=pypi_0
itsdangerous=2.2.0=pypi_0
jax=0.4.30=pypi_0
jaxlib=0.4.30=pypi_0
jinja2=3.1.4=pypi_0
jiter=0.5.0=pypi_0
joblib=1.4.2=pypi_0
keras=3.4.1=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.3=pypi_0
kornia-rs=0.1.5=pypi_0
lazy-loader=0.4=pypi_0
libblas=3.9.0=23_win64_mkl
libcblas=3.9.0=23_win64_mkl
libclang=18.1.1=pypi_0
libffi=3.4.2=h8ffe710_5
libhwloc=2.11.1=default_h8125262_1000
libiconv=1.17=hcfcfb64_2
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=23_win64_mkl
libpng=1.6.43=h19919ed_0
librosa=0.10.2.post1=pypi_0
libsqlite=3.46.0=h2466b09_0
libxml2=2.12.7=h0f24e4e_4
libzlib=1.3.1=h2466b09_1
libzlib-wapi=1.2.13=h2466b09_6
llvmlite=0.43.0=pypi_0
lxml=4.9.4=pypi_0
mako=1.3.5=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.1=pypi_0
mdurl=0.1.2=pypi_0
mediapipe=0.10.14=pypi_0
mkl=2021.4.0=pypi_0
ml-dtypes=0.4.0=pypi_0
mmengine=0.10.4=pypi_0
more-itertools=10.3.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
mtcnn=0.1.1=pypi_0
mysqlclient=2.2.4=pypi_0
namex=0.0.8=pypi_0
networkx=3.2.1=pypi_0
nibabel=5.2.1=pypi_0
nilearn=0.10.4=pypi_0
nltk=3.8.1=pypi_0
nltools=0.5.1=pypi_0
numba=0.60.0=pypi_0
numexpr=2.8.4=pypi_0
numpy=1.23.5=pypi_0
openai=1.37.0=pypi_0
openai-whisper=20231117=pypi_0
opencv-contrib-python=4.10.0.84=pypi_0
opencv-python=4.10.0.84=pypi_0
opensmile=2.5.0=pypi_0
openssl=3.3.1=h2466b09_2
opt-einsum=3.3.0=pypi_0
optree=0.12.1=pypi_0
orjson=3.10.6=pypi_0
oyaml=1.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.0=py39haa95532_0
platformdirs=4.2.2=pypi_0
pooch=1.8.2=pypi_0
praat-parselmouth=0.4.3=pypi_0
proglog=0.1.10=pypi_0
protobuf=4.25.3=pypi_0
pthreads-win32=2.9.1=hfa6e2cd_3
py-feat=0.6.2=pypi_0
pyarrow=17.0.0=pypi_0
pyaudio=0.2.14=pypi_0
pycparser=2.22=pypi_0
pycryptodomex=3.20.0=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pypi_0
pyjwt=2.8.0=pypi_0
pynv=0.3=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.19=h4de0772_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-decouple=3.8=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
python-speech-features=0.6=pypi_0
python_abi=3.9=4_cp39
pytz=2024.1=pypi_0
pywavelets=1.6.0=pypi_0
pyyaml=6.0.1=pypi_0
regex=2024.5.15=pypi_0
reportlab=4.2.2=pypi_0
requests=2.32.3=pypi_0
rest-framework-simplejwt=0.0.2=pypi_0
retina-face=0.0.17=pypi_0
rich=13.7.1=pypi_0
ruff=0.5.5=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=69.5.1=py39haa95532_0
shellingham=1.5.4=pypi_0
simplejwt=2.0.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sounddevice=0.4.7=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.5=pypi_0
soxr=0.3.7=pypi_0
sqlalchemy=1.4.53=pypi_0
sqlite=3.45.3=h2bbff1b_0
sqlparse=0.5.1=pypi_0
starlette=0.37.2=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.13.0=pypi_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.17.0=pypi_0
tensorflow-intel=2.17.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
tf-keras=2.17.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.7.24=pypi_0
tiktoken=0.4.0=pypi_0
timm=1.0.7=pypi_0
tk=8.6.13=h5226925_1
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
tomlkit=0.12.0=pypi_0
torch=2.3.1=pypi_0
torchaudio=2.3.1=pypi_0
torchvision=0.18.1=pypi_0
tqdm=4.66.4=pypi_0
transformers=4.43.3=pypi_0
typer=0.12.3=pypi_0
typing=3.7.4.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.3=pypi_0
vadersentiment=3.3.2=pypi_0
vc=14.2=h2eaa2aa_4
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
watchfiles=0.22.0=pypi_0
webrtcvad=2.0.10=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39haa95532_0
whisper=1.1.10=pypi_0
wordcloud=1.9.3=pypi_0
wrapt=1.16.0=pypi_0
xgboost=2.1.1=pypi_0
xz=5.2.6=h8d14728_0
yapf=0.40.2=pypi_0
zipp=3.19.2=pypi_0
